+++ addon/components/flexberry-lookup.js
@@ -307,7 +307,8 @@ export default FlexberryBaseComponent.extend({
       if (lookupWindowCustomProperties) {
         let result = lookupWindowCustomProperties({
           relationName: this.get('relationName'),
-          projection: this.get('projection')
+          projection: this.get('projection'),
+          componentName: this.get('componentName')
         });
 
         return result;
@@ -602,7 +603,7 @@ export default FlexberryBaseComponent.extend({
     @method didDestroyElement
   */
   didDestroyElement() {
-    this._super();
+    this._super(...arguments);
     this.removeObserver('i18n.locale', this, this._languageReinit);
   },
 
@@ -613,7 +614,7 @@ export default FlexberryBaseComponent.extend({
     @method didInsertElement
   */
   didInsertElement() {
-    this._super();
+    this._super(...arguments);
     this.addObserver('i18n.locale', this, this._languageReinit);
   },
 
@@ -624,7 +625,7 @@ export default FlexberryBaseComponent.extend({
     @method didRender
   */
   didRender() {
-    this._super();
+    this._super(...arguments);
 
     let isAutocomplete = this.get('autocomplete');
     let isDropdown = this.get('dropdown');