+++ blueprints/flexberry-model/index.js
@@ -64,7 +64,8 @@ module.exports = {
             name: modelBlueprint.name,
             needsAllModels: modelBlueprint.needsAllModels,
             needsAllEnums: modelBlueprint.needsAllEnums,
-            needsAllObjects: modelBlueprint.needsAllObjects // for use in files\tests\unit\serializers\__name__.js
+            needsAllObjects: modelBlueprint.needsAllObjects,
+            enumImports: modelBlueprint.enumImports,
         }, modelBlueprint.lodashVariables);
     }
 };